完成借阅信息service类,修改了主键Long的相关问题，修改了时间的相 关问题

+++ manager_springboot/src/main/java/com/gmy/mapper/UserMapper.java
@@ -17,6 +17,6 @@ import org.springframework.stereotype.Repository;
 public interface UserMapper extends BaseMapper<User> {
 
     @Select("select id from user where ACCOUNT=#{account}")
-    public int getId(String account);
+    public Long getId(String account);
 
 }
